refactor: user, user_repository 파일 이동 및 user, pet 클래스 분리
